Split type generator
